requirements.txt
numpy>=1.21
transformers>=4.22
wandb>=0.13
panphon>=0.20
epitran>=1.20
tqdm>=4.64
scikit-learn>=1.1
scipy>=1.7